refactor(emil): rename email.py to email_utils.py
